Create jump game
Given an array of non-negative integers, you are initially positioned at the first index of the array.

Each element in the array represents your maximum jump length at that position.

Determine if you are able to reach the last index.

For example:
A = [2,3,1,1,4], return true.

A = [3,2,1,0,4], return false. 

idea: we start from the second-rightmost position.
the basic idea is if we can reach the next position that leads to the final index, then we can reach the final index position.
if not we move to the next position from right to left and increment required distance by 1;
if we can reach, we reset distance be 1.
for instance: [2,3,1,1,4] we start from 3rd position(1): since 1 is >= 1 then we move to next 2nd position(1): since 1 is >= 1 and we repeat until we reach the leftmost position and return True;
another example: [3,2,1,0,4] we start from 3rd position(0): since 0 is < initial distance 1 then we move to 2nd position(1) and set distance be 2, since 2nd position (1) is < 2 then again we move to 1st position(2) and set distance be 3. Until we reach the leftmost position and realize that 3 is less than required distance 4. Thus return False.