Create 16 3sum closest
Given an array S of n integers, find three integers in S such that the sum is closest to a given number, target. Return the sum of the three integers. You may assume that each input would have exactly one solution.

    For example, given array S = {-1 2 1 -4}, and target = 1.

    The sum that is closest to the target is 2. (-1 + 2 + 1 = 2).

idea: the idea is kind of similar to 3sum problem:
1) sort input array
2) we fix the first member from the list by traversing the input array from [0..N-3]
3) we let second and third member starting from 0 and N-1 respectively:
the most possibile situation that would give us the closest sum to target is the last sum before the relation between target and threesum begin changing:
ex: if sum is 98 at some point and target is 95, since 98 is bigger than 95, we forward third member and we have 90, since 90 is less than 95 then we have to advance second member, thus as you can see 98 is what we called as the most possible situation that would give us the closest result to target ! (we call the most possible situation as turning point as comments in the above code !)